requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
datasets==2.14.6
dill==0.3.7
distro==1.9.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djongo==1.3.7
docstring_parser==0.17.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2023.10.0
google-auth==2.48.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.15
networkx==3.6.1
nltk==3.9.2
numpy==1.26.4
ollama==0.6.1
packaging==26.0
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
propcache==0.4.1
psutil==7.2.2
pyarrow==14.0.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==3.11.4
PyPDF2==3.0.1
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==5.0.0
typer==0.24.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
zss==1.2.0
